Consolidate people in address book
CommandTestUtil and TypicalPersons used 2 different groups of people for different tests in the test suite, making it hard to keep track of the details of the people used and can cause some regressions when changes are made.

Consolidating this group into CommandTestUtil would allow for a single source of information. Staff IDs and passport numbers of these people are taken from TypicalStaffIds and TypicalPassportNumbers