Change use cases and fix numbering
Split use cases to handle a single entity. For example, one use case for editing a staff and one for a guest.

Also fixed the numbering for extensions in use cases. Extensions should be the same number as the situation that occurs prior to it happening.
For example if a and b occur in MSS (numbered 1 and 2). Instead of b occuring, c can occurs (c is the extension). c should be numbered 1 instead of 2